refactor __init__.py, add example plugin
